requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
executing==2.2.0
fonttools==4.58.4
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.66.2
idna==3.10
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==2.3.0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pbr==6.1.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.2.3
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==27.0.0
qiskit==2.1.0
qiskit-aer==0.17.1
qiskit-ibm-runtime==0.40.1
requests==2.32.4
requests_ntlm==1.3.0
rustworkx==0.16.0
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
stevedore==5.4.1
sympy==1.14.0
tornado==6.5.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
